So when I tried calling mkdir on the client.py stuff, it threw errors telling me that OneDirFptClient has no attribute "mkdir," and I'm not sure why that's happening. I've got it printing out the parameters it will need to update the server, and the calls to the server for creation and renaming are in the code, but commented out for the time being.
I'm not sure why the renaming is having so many problems. For some reason, the rename doesn't get detected, but it instead goes to the default "Rename or Modified!" but once it gets there, it doesn't pick up the new file name. If we could trigger the renaming, I think we could fix the problem.